qemu/main: Make GC heap size configurable on a per-arch basis.
This commit introduces a new configuration variable for having different
GC heap sizes depending on the architecture this port will be built for.

This is needed to fix two issues: on RV32, in certain cases that depend
on the final code size of the binary to be run, the `deflate_decompress`
test may fail due to an unsuccessful memory allocation, and on RV64 the
`memstats` test fails due to an unsuccessful memory allocation.

In the first case, bumping up the GC heap size by 20KBytes (from 100 to
120) solves the issue, whilst the second case requires a bit more than
that (being on a 64 bits platform, perhaps) and thus the GC heap size is
doubled (from 100 to 200).

AArch32 ports do not require further tweaks, and their GC heap size is
left to the default value.

Signed-off-by: Alessandro Gatti <a.gatti@frob.it>